S#007

+++ bin/fix_ssh_agent.sh
@@ -3,52 +3,73 @@
 
 # INIT {{{
 if [[ $1 == '@@' ]]; then
-  echo "fix_ssh_agent start_if_needed fix_tmux"
+  echo "fix_ssh_agent start_if_needed fix_tmux --help"
   exit 0
 fi
 
-[[ "$BASH_SOURCE" == "$0" ]] && echo "Script has to be sourced!" && exit 1
+
+if [[ "$BASH_SOURCE" == "$0" ]]; then
+  echo "Script has to be sourced!"
+  echo
+  echo "Usage:"
+  echo "  source $0 [fix_ssh_agent | start_if_needed | fix_tmux]"
+  echo "source $0 start_if_needed" | ccopy
+  exit 1
+fi
 
 SSH_PATH="$TMP_MEM_PATH/.ssh/$(hostname)"
 SSH_ENV="$SSH_PATH/environment"
 LINK_PATH="$SSH_PATH/ssh_auth_sock"
 # }}}
 # Functions {{{
-fix_ssh_agent() { # {{{
-  [[ -z $SSH_AUTH_SOCK ]] && return 0
-  [[ ! -e $SSH_PATH ]] && command mkdir -p $SSH_PATH
-  [[ -e $LINK_PATH && $SSH_AUTH_SOCK == $LINK_PATH ]] && return 0
-  [[ $SSH_AUTH_SOCK == $LINK_PATH ]] && return 1
-  [[ ! -e $LINK_PATH ]] && ln -sf $SSH_AUTH_SOCK $LINK_PATH
-  chmod 600 $LINK_PATH 2>/dev/null
-  export SSH_AUTH_SOCK=$LINK_PATH
-} # }}}
 fix_tmux() { # {{{
   [[ -z $SSH_AUTH_SOCK ]] && return 0
   tmux set-environment -g 'SSH_AUTH_SOCK' $LINK_PATH
 } # }}}
-start_agent() { # {{{
-  local stderr="$SSH_PATH/stderr"
-  ssh-agent -a "$SSH_PATH/bind" -s 2>$stderr | sed 's/^echo/#echo/' >$SSH_ENV
-  if [[ -s $stderr ]]; then
-    echo "Failed to start ssh-agent !" >/dev/stderr
-    rm $SSH_ENV
-  else
-    chmod 600 $SSH_ENV
-    source $SSH_ENV > /dev/null
+fix_ssh_agent() { # {{{
+  [[ ! -e $SSH_PATH ]] && return 1
+  if [[ -z $SSH_AUTH_SOCK ]]; then
+    [[ -f $SSH_ENV ]] && source $SSH_ENV >/dev/null
+    [[ -z $SSH_AUTH_SOCK ]] && return 1
   fi
+  [[ $SSH_AUTH_SOCK == $LINK_PATH && -e $LINK_PATH ]] && return 0
+  [[ -e $(readlink -f $LINK_PATH) ]] && SSH_AUTH_SOCK=$LINK_PATH && return 0
+  rm -f $LINK_PATH
+  ln -sf $SSH_AUTH_SOCK $LINK_PATH
+  chmod 600 $LINK_PATH 2>/dev/null
+  export SSH_AUTH_SOCK=$LINK_PATH
+} # }}}
+start_agent() { # {{{
+  local err=0 i=0 stderr="$SSH_PATH/stderr"
+  while [[ $i -le 1 ]]; do
+   command mkdir -p $SSH_PATH && chmod 700 $SSH_PATH
+    if ssh-agent -a "$SSH_PATH/bind" -s 2>$stderr | sed 's/^echo/#echo/' >$SSH_ENV && [[ ! -s $stderr ]]; then
+      chmod 600 $SSH_ENV
+      source $SSH_ENV >/dev/null
+      err=0; break
+    else
+      rm $SSH_ENV
+      err=$?
+    fi
+    command rm -rf $SSH_PATH
+    i=$(($i+1))
+  done
   rm -f $stderr
+  if [[ $err != 0 ]]; then
+    echo "Failed to start ssh-agent ($err)" >/dev/stderr
+  fi
+  return $err
 } # }}}
 start_if_needed() { # {{{
   [[ -n "$SSH_CLIENT" || -n "$SSH_TTY" ]] && return 0
-  [[ ! -e $SSH_PATH ]] && command mkdir -p $SSH_PATH
-  chmod 700 $SSH_PATH
   local start_ssh=true
   if [[ -f $SSH_ENV ]]; then
-    source $SSH_ENV > /dev/null
-    [[ ! -z $SSH_AGENT_PID ]] && command ps -ef | command grep -e $SSH_AGENT_PID | command grep -qe 'ssh-agent' && start_ssh=false
+    source $SSH_ENV >/dev/null
+    [[ ! -z $SSH_AGENT_PID ]] && pidof ssh-agent | grep -q "\<$SSH_AGENT_PID\>" && start_ssh=false
   fi
-  $start_ssh && start_agent
+  if $start_ssh; then
+    start_agent || return 1
+  fi
   fix_ssh_agent
 } # }}}
 # }}}
@@ -60,13 +81,7 @@ fix_ssh_agent|start_if_needed|fix_tmux) $cmd $@;;
 esac
 # }}}
 # Cleaning {{{
-unset cmd
-unset SSH_PATH
-unset SSH_ENV
-unset LINK_PATH
-unset fix_ssh_agent
-unset fix_tmux
-unset start_if_needed
-unset start_agent
+unset cmd SSH_PATH SSH_ENV LINK_PATH
+unset fix_ssh_agent fix_tmux start_if_needed start_agent
 # }}}
